WIP - correcting and assuring floating point data type in conversion formula.
