+++ drivers/kionix/kx132-1211/kx132-conversions.h
@@ -19,8 +19,8 @@
 #ifndef KX132_ACCELEROMETER_RANGE_DEFINES
 #define KX132_ACCELEROMETER_RANGE_DEFINES
 
-#define KX132_RANGE_RES_HIGH_DECIMAL_COUNTS_MIN -32768
-#define KX132_RANGE_RES_HIGH_DECIMAL_COUNTS_MAX  32767
+#define KX132_RANGE_RES_HIGH_DECIMAL_COUNTS_MIN -32768.0
+#define KX132_RANGE_RES_HIGH_DECIMAL_COUNTS_MAX  32767.0
 
 #define KX132_RANGE_RES_HIGH_2G_MIN            -2.00000
 #define KX132_RANGE_RES_HIGH_2G_MAX             1.99994
@@ -33,25 +33,6 @@
 
 #define KX132_RANGE_RES_HIGH_16G_MIN          -16.00000
 #define KX132_RANGE_RES_HIGH_16G_MAX           15.99951
-
-// 12-7 NEED to add low resolution range values (these for 8-bit readings).
-
-
-enum kx132_acceleration_ranges
-{
-    KX132_ACCEL_RANGES_BEGIN,
-    KX132_RANGE_PLUS_MINUS_2G,
-    KX132_RANGE_PLUS_MINUS_4G,
-    KX132_RANGE_PLUS_MINUS_8G,
-    KX132_RANGE_PLUS_MINUS_16G,
-    KX132_ACCEL_RANGES_END
-};
-
-enum kx132_acceleration_resolutions
-{
-    KX132_ACCEL_RESOLUTION_HIGH = 1,
-    KX132_ACCEL_RESOLUTION_LOW
-};
 
 #endif // KX132_ACCELEROMETER_RANGE_DEFINES
